Fixed broken vital strike
