+++ Repository.json
@@ -2,7 +2,7 @@
     "Releases": [
         {
             "Id": "TabletopTweaks",
-            "Version": "1.0.6"
+            "Version": "1.0.7"
         }
     ]
 }
